Posting a user

+++ SecretSanta/test/SecretSanta.Web.Tests/GiftTests.cs
@@ -10,8 +10,8 @@ using System.Linq;
 using System.Diagnostics;
 using SecretSanta.Web.Api;
 using System.Net.Http;
-using System.Text.Json;
-using System.Text;
+using System.Threading.Tasks;
+using System.Net.Http.Headers;
 
 namespace SecretSanta.Web.Tests
 {
@@ -22,69 +22,61 @@ namespace SecretSanta.Web.Tests
         public TestContext? TestContext { get; set; }
         [NotNull]
         private IWebDriver? Driver { get; set; }
-        private static string ApiUrl { get;} = "https://localhost:44388";
-        private static string WebUrl { get; } = "https://localhost:44394";
         private static Process? ApiHostProcess { get; set; }
-        private static Process? NpmProcess { get; set; }
         private static Process? WebHostProcess { get; set; }
         string AppUrl { get; } = "https://localhost:44394/Gifts";
 
         [ClassInitialize]
-        public static void ClassInitalize(TestContext testContext)
+        public static async Task ClassInitalize(TestContext testContext)
         {
+            if (testContext is null)
+                throw new ArgumentNullException(nameof(testContext));
 
-            //var psiNpmRunDist = new ProcessStartInfo
-            //{
-            //    FileName = "cmd",
-            //    RedirectStandardInput = true,
-            //    WorkingDirectory = "C:\\Users\\mason\\Source\\Repos\\EWU-CSCD379-2020-Winter\\SecretSanta\\src\\SecretSanta.Web"
-            //};
-            //NpmProcess = Process.Start(psiNpmRunDist);
-            //NpmProcess.StandardInput.WriteLine("npm run build:dev");
-            //NpmProcess.WaitForExit(15000);
-            //NpmProcess = Process.Start("cmd", "npm run build:dev run --prefix ..\\..\\..\\..\\..\\src\\SecretSanta.Web");
-            //UserClient userClient = new UserClient();
-            HttpClient client = new HttpClient();   
-            //Uri uri = new Uri(ApiUrl + "api/User", UriKind.RelativeOrAbsolute);
-            User user = new User();
-            user.FirstName = "inigo";
-            user.LastName = "montoya";
-            //UserClient userClient = new UserClient(new HttpClient());
-            //userClient.PostAsync(user);
-            //Console.WriteLine(user.Id);
-            using var content = new StringContent(JsonSerializer.Serialize(user), Encoding.UTF8, "application/json");
-            client.PostAsync(ApiUrl+ "/api/User", content);
-            client.PostAsync(ApiUrl, content);
             string ProjectPath = testContext.DeploymentDirectory;
-            
-            string ApiProjectPath = ProjectPath + "\\src\\SecretSanta.Api";
-            string WebProjectPath = ProjectPath + "\\src\\SecretSanta.Web";
+
             Console.WriteLine(ProjectPath);
             ApiHostProcess = Process.Start("dotnet.exe", "run -p ..\\..\\..\\..\\..\\src\\SecretSanta.Api\\SecretSanta.Api.csproj");
             WebHostProcess = Process.Start("dotnet.exe", "run -p ..\\..\\..\\..\\..\\src\\SecretSanta.Web\\SecretSanta.Web.csproj");
-            //UserClient userClient = new UserClient(new HttpClient());
-            //userClient.PostAsync(user);
-            
+
+            //AddUser
+            HttpClient httpClient = new HttpClient();
+            httpClient.BaseAddress = new Uri("https://localhost:44388/");
+            httpClient.DefaultRequestHeaders.Accept.Clear();
+            httpClient.DefaultRequestHeaders.Accept.Add(
+                new MediaTypeWithQualityHeaderValue("application/json"));
+
+            UserClient userClient = new UserClient(httpClient);
+            UserInput userInput = new UserInput();
+
+            userInput.FirstName = "Test";
+            userInput.LastName = "User";
+            await userClient.PostAsync(userInput);
+            httpClient.Dispose();
 
         }
+
         [ClassCleanup]
         public static void ClassCleanup()
         {
-            //NpmProcess.Kill();
-            //NpmProcess.Close();
-            //NpmProcess.CloseMainWindow();
-            ApiHostProcess.Kill();
-            ApiHostProcess.CloseMainWindow();
-            ApiHostProcess.Close();
-            WebHostProcess.Kill();
-            WebHostProcess.CloseMainWindow();
-            WebHostProcess.Close();
+            if (ApiHostProcess != null)
+            {
+                ApiHostProcess.Kill();
+                ApiHostProcess.CloseMainWindow();
+                ApiHostProcess.Close();
+            }
+            if (WebHostProcess != null)
+            {
+                WebHostProcess.Kill();
+                WebHostProcess.CloseMainWindow();
+                WebHostProcess.Close();
+            }
         }
+
         [TestInitialize]
         public void TestInitialize()
         {
-            //Driver = new ChromeDriver();
-            Driver = new ChromeDriver(Environment.GetEnvironmentVariable("ChromeWebDriver"));//supposedly required for Azure
+            Driver = new ChromeDriver();
+            //Driver = new ChromeDriver(Environment.GetEnvironmentVariable("ChromeWebDriver"));//supposedly required for Azure
             Driver.Manage().Timeouts().ImplicitWait = new System.TimeSpan(0, 0, 10);
         }
 
@@ -118,8 +110,8 @@ namespace SecretSanta.Web.Tests
 
             //Assert
             Assert.AreEqual<string>(title, giftTitleElement.GetProperty("value"));
-           Assert.AreEqual<string>(description, giftDescriptionElement.GetProperty("value"));
-           Assert.AreEqual<string>(url, giftUrlElement.GetProperty("value"));
+            Assert.AreEqual<string>(description, giftDescriptionElement.GetProperty("value"));
+            Assert.AreEqual<string>(url, giftUrlElement.GetProperty("value"));
         }
 
         public void ClickSubmitButton()
@@ -144,11 +136,8 @@ namespace SecretSanta.Web.Tests
             string title = "TestGift",
                    description = "This gift was made by the 'Create_Gift_Success()' test.",
                    url = "www.gifttest.com";
-            Thread.Sleep(2000);
             ClickCreateButton();
-            Thread.Sleep(5000);
             EnterGiftInformation(title, description, url, 0);
-            Thread.Sleep(5000);
             ClickSubmitButton();
             Thread.Sleep(5000);
 
@@ -157,9 +146,9 @@ namespace SecretSanta.Web.Tests
             int index = giftAttributes.IndexOf(title);
 
             //Assert
-            Assert.AreEqual(title,giftAttributes[index]);
-            Assert.AreEqual(description, giftAttributes[index+1]);
-            Assert.AreEqual(url, giftAttributes[index+2]);
+            Assert.AreEqual(title, giftAttributes[index]);
+            Assert.AreEqual(description, giftAttributes[index + 1]);
+            Assert.AreEqual(url, giftAttributes[index + 2]);
 
             //Take screen shot upon success
             TakeScreenShot("Create_Gift_Success_Test_Screenshot");
